Enable for nested converters.
Since now converters that are able to convert from upper classes are used in case if there's no strict converter provided.
Assuming that class A extends B

Converter<B,C>

Will be used when convert(a, c).
But if registered converters
Converter<A,C>
Converter<B,C>
Then in case convert(a,c) will be used Converter<A,C>